+++ AddRemoveCustomProperties/release_notes.md
@@ -1,1 +1,7 @@
 ## v0.1.0 Initial release
+
+## v2024.1 
+
+Added the option to remove properties that ARE NOT on the list in `property_list.txt`.  Previous behavior was to always remove properties that ARE on the list.
+
+Fixed an issue where the old error handling method was still being used.